Add referral type dropdown with only one option
