+++ src/declarations.d.tsx
@@ -3,3 +3,16 @@
 declare type Observable<T> = import("rxjs").Observable<T>;
 
 declare module "*.css";
+
+type Referral = {
+  details: string;
+  encounter_uuid: string;
+  encounter_id: number;
+  person_uuid: string;
+  patient_name: string;
+  referral_date: string;
+  referral_type: string;
+  status: string;
+  visit_uuid: string;
+  zl_emr_id: string;
+};
